+++ src/main/java/me/drex/worldmanager/command/DeleteCommand.java
@@ -2,11 +2,11 @@ package me.drex.worldmanager.command;
 
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
-import me.drex.message.api.LocalizedMessage;
 import me.drex.worldmanager.save.WorldManagerSavedData;
 import me.lucko.fabric.api.permissions.v0.Permissions;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.commands.arguments.ResourceLocationArgument;
+import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 
 import static me.drex.worldmanager.command.WorldManagerCommand.CUSTOM_WORLD_SUGGESTIONS;
@@ -32,7 +32,7 @@ public class DeleteCommand {
         if (!success) {
             throw UNKNOWN_WORLD.create();
         }
-        source.sendSuccess(() -> LocalizedMessage.builder("worldmanager.command.delete").addPlaceholder("id", id.toString()).build(), false);
+        source.sendSuccess(() -> Component.literal("World " + id + " has been deleted successfully!"), false);
         return 1;
     }
 }