+++ src/main/java/me/drex/worldmanager/command/TeleportCommand.java
@@ -14,6 +14,7 @@ import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.commands.arguments.EntityArgument;
 import net.minecraft.commands.arguments.ResourceLocationArgument;
 import net.minecraft.core.registries.Registries;
+import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.MinecraftServer;
@@ -25,7 +26,6 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
 
-import static me.drex.message.api.LocalizedMessage.builder;
 import static me.drex.worldmanager.command.WorldManagerCommand.UNKNOWN_WORLD;
 import static me.drex.worldmanager.command.WorldManagerCommand.WORLD_SUGGESTIONS;
 import static net.minecraft.commands.Commands.argument;
@@ -58,7 +58,7 @@ public class TeleportCommand {
         for (ServerPlayer player : targets) {
             teleport(player, config, id);
         }
-        source.sendSuccess(() -> builder("worldmanager.command.teleport").addPlaceholder("id", id.toString()).build(), false);
+        source.sendSuccess(() -> Component.literal("Teleporting to " + id), false);
         return 1;
     }