+++ src/main/java/me/drex/worldmanager/gui/configure/ConfigureWorld.java
@@ -2,34 +2,23 @@ package me.drex.worldmanager.gui.configure;
 
 import eu.pb4.sgui.api.elements.GuiElementBuilder;
 import eu.pb4.sgui.api.gui.SimpleGui;
-import me.drex.message.api.LocalizedMessage;
 import me.drex.worldmanager.save.WorldConfig;
 import me.drex.worldmanager.save.WorldData;
-import me.drex.worldmanager.save.WorldManagerSavedData;
+import net.minecraft.ChatFormatting;
 import net.minecraft.core.Holder;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.core.RegistryAccess;
 import net.minecraft.core.registries.Registries;
-import net.minecraft.nbt.CompoundTag;
-import net.minecraft.nbt.NbtOps;
-import net.minecraft.nbt.Tag;
-import net.minecraft.nbt.TextComponentTagVisitor;
 import net.minecraft.network.chat.Component;
-import net.minecraft.resources.RegistryOps;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.inventory.MenuType;
-import net.minecraft.world.item.Item;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.level.chunk.ChunkGenerator;
 import net.minecraft.world.level.dimension.BuiltinDimensionTypes;
 import net.minecraft.world.level.dimension.DimensionType;
-import xyz.nucleoid.fantasy.Fantasy;
-import xyz.nucleoid.fantasy.RuntimeWorldHandle;
-
-import java.util.Optional;
 
 public abstract class ConfigureWorld extends SimpleGui {
 
@@ -42,7 +31,7 @@ public abstract class ConfigureWorld extends SimpleGui {
     public ConfigureWorld(ServerPlayer player, ResourceLocation id) {
         super(MenuType.GENERIC_9x3, player, false);
         this.id = id;
-        setTitle(LocalizedMessage.builder("worldmanager.gui.configure.title").addPlaceholder("id", id.toString()).build());
+        setTitle(Component.literal("Configure World " + id));
         setupDefaults(player.level().getServer());
         build();
     }
@@ -58,7 +47,10 @@ public abstract class ConfigureWorld extends SimpleGui {
 
     public void build() {
         setSlot(0,
-            builder(Items.ARMOR_STAND, "type")
+            new GuiElementBuilder(Items.ARMOR_STAND)
+                .setName(Component.literal("Dimension Type").withStyle(ChatFormatting.GREEN))
+                .addLoreLine(Component.literal("World height, sky light, beds working, ...").withStyle(ChatFormatting.GRAY))
+                .addLoreLine(Component.literal("Current: ").withStyle(ChatFormatting.GRAY).append(Component.literal(type.unwrapKey().map(ResourceKey::location).map(ResourceLocation::toString).orElse("???")).withStyle(ChatFormatting.GOLD)))
                 .setCallback(() -> {
                     new SelectDimensionType(player, this, type -> {
                         this.type = type;
@@ -67,7 +59,10 @@ public abstract class ConfigureWorld extends SimpleGui {
                 })
         );
         setSlot(1,
-            builder(Items.GRASS_BLOCK, "generator")
+            new GuiElementBuilder(Items.GRASS_BLOCK)
+                .setName(Component.literal("Chunk Generator").withStyle(ChatFormatting.GREEN))
+                .addLoreLine(Component.literal("How the world is shaped").withStyle(ChatFormatting.GRAY))
+                .addLoreLine(Component.literal("Current: ").withStyle(ChatFormatting.GRAY).append(Component.literal(generator.getTypeNameForDataFixer().map(ResourceKey::location).map(ResourceLocation::toString).orElse("???")).withStyle(ChatFormatting.GOLD)))
                 .setCallback(() -> {
                     new SelectChunkGenerator(player, this, generator -> {
                         this.generator = generator;
@@ -76,7 +71,9 @@ public abstract class ConfigureWorld extends SimpleGui {
                 })
         );
         setSlot(2,
-            builder(Items.WHEAT_SEEDS, "seed")
+            new GuiElementBuilder(Items.WHEAT_SEEDS)
+                .setName(Component.literal("World Seed").withStyle(ChatFormatting.LIGHT_PURPLE))
+                .addLoreLine(Component.literal("Current: ").withStyle(ChatFormatting.GRAY).append(Component.literal(String.valueOf(seed)).withStyle(ChatFormatting.GOLD)))
                 .setCallback(() -> {
                     new SelectSeed(player, this, seed -> {
                         this.seed = seed;
@@ -85,7 +82,10 @@ public abstract class ConfigureWorld extends SimpleGui {
                 })
         );
         setSlot(3,
-            builder(Items.CLOCK, "tick_time")
+            new GuiElementBuilder(Items.CLOCK)
+                .setName(Component.literal("Time Ticking").withStyle(ChatFormatting.AQUA))
+                .addLoreLine(Component.literal("Enable / disable daylight cycle").withStyle(ChatFormatting.GRAY))
+                .addLoreLine(Component.literal("Current: ").withStyle(ChatFormatting.GRAY).append(Component.literal(tickTime ? "Enabled" : "Disabled").withStyle(tickTime ? ChatFormatting.GREEN : ChatFormatting.RED)))
                 .setCallback(() -> {
                     this.tickTime = !this.tickTime;
                     build();
@@ -94,7 +94,8 @@ public abstract class ConfigureWorld extends SimpleGui {
 
         for (int i = 2 * 9; i < 3 * 9; i++) {
             setSlot(i,
-                builder(Items.GREEN_STAINED_GLASS_PANE, "confirm")
+                new GuiElementBuilder(Items.GREEN_STAINED_GLASS_PANE)
+                    .setName(Component.literal("Confirm!").withStyle(ChatFormatting.GREEN, ChatFormatting.BOLD))
                     .setCallback(() -> {
                         WorldConfig config = new WorldConfig(
                             type,
@@ -111,23 +112,4 @@ public abstract class ConfigureWorld extends SimpleGui {
     }
 
     protected abstract void confirm(WorldConfig config);
-
-    private GuiElementBuilder builder(Item item, String id) {
-        var access = player.registryAccess();
-        Optional<Tag> result = ChunkGenerator.CODEC.encodeStart(RegistryOps.create(NbtOps.INSTANCE, access), generator).result();
-        Component lore = Component.empty();
-        if (result.isPresent()) {
-            lore = new TextComponentTagVisitor("    ").visit(result.get());
-        }
-        return new GuiElementBuilder(item)
-            .setName(LocalizedMessage.builder("worldmanager.gui.configure." + id + ".name")
-                .build())
-            .addLoreLine(LocalizedMessage.builder("worldmanager.gui.configure." + id + ".lore")
-                .addPlaceholder("type", type.unwrapKey().map(ResourceKey::location).map(ResourceLocation::toString).orElse("???"))
-                .addPlaceholder("generator", generator.getTypeNameForDataFixer().map(ResourceKey::location).map(ResourceLocation::toString).orElse("???"))
-                .addPlaceholder("generator_full", lore)
-                .addPlaceholder("seed", seed)
-                .addPlaceholder("tick_time", tickTime)
-                .build());
-    }
 }