+++ src/main/java/me/drex/worldmanager/gui/list/ManageWorld.java
@@ -2,7 +2,6 @@ package me.drex.worldmanager.gui.list;
 
 import eu.pb4.sgui.api.elements.GuiElementBuilder;
 import eu.pb4.sgui.api.gui.SimpleGui;
-import me.drex.message.api.LocalizedMessage;
 import me.drex.worldmanager.command.SpawnCommand;
 import me.drex.worldmanager.command.TeleportCommand;
 import me.drex.worldmanager.gui.util.ConfirmTypeGui;
@@ -10,14 +9,13 @@ import me.drex.worldmanager.gui.util.GuiElements;
 import me.drex.worldmanager.save.WorldConfig;
 import me.drex.worldmanager.save.WorldManagerSavedData;
 import me.lucko.fabric.api.permissions.v0.Permissions;
+import net.minecraft.ChatFormatting;
+import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.inventory.MenuType;
-import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
-
-import static me.drex.message.api.LocalizedMessage.builder;
 
 public class ManageWorld extends SimpleGui {
     private final ResourceLocation id;
@@ -29,13 +27,15 @@ public class ManageWorld extends SimpleGui {
         this.id = id;
         this.config = config;
         this.previousGui = previousGui;
-        setTitle(builder("worldmanager.gui.manage_world.title").addPlaceholder("id", id.toString()).build());
+        setTitle(Component.literal("Manage World " + id));
         build();
     }
 
     public void build() {
         setSlot(0,
-            guiElementBuilder(Items.ENDER_PEARL, "teleport")
+            new GuiElementBuilder(Items.ENDER_PEARL)
+                .setName(Component.literal("Teleport Back").withStyle(ChatFormatting.AQUA))
+                .addLoreLine(Component.literal("Teleport to your last location in the world.").withStyle(ChatFormatting.GRAY))
                 .setCallback(() -> {
                     if (Permissions.check(player, "worldmanager.command.worldmanager.teleport", 2)) {
                         if (TeleportCommand.teleport(player, config, id)) {
@@ -45,7 +45,9 @@ public class ManageWorld extends SimpleGui {
                 })
         );
         setSlot(1,
-            guiElementBuilder(Items.GRASS_BLOCK, "spawn")
+            new GuiElementBuilder(Items.GRASS_BLOCK)
+                .setName(Component.literal("Teleport To Spawn").withStyle(ChatFormatting.GREEN))
+                .addLoreLine(Component.literal("Teleport to the world spawn.").withStyle(ChatFormatting.GRAY))
                 .setCallback(() -> {
                     if (Permissions.check(player, "worldmanager.command.worldmanager.spawn", 2)) {
                         if (SpawnCommand.spawn(player, config, id)) {
@@ -55,7 +57,9 @@ public class ManageWorld extends SimpleGui {
                 })
         );
         setSlot(2,
-            guiElementBuilder(config.data.iconGuiElement(), "icon")
+            config.data.iconGuiElement()
+                .setName(Component.literal("Change World Icon").withStyle(ChatFormatting.YELLOW))
+                .addLoreLine(Component.literal("Set the world icon to the item in your hand.").withStyle(ChatFormatting.GRAY))
                 .setCallback(() -> {
                     if (Permissions.check(player, "worldmanager.command.worldmanager.seticon", 2)) {
                         ItemStack item = player.getMainHandItem();
@@ -69,7 +73,9 @@ public class ManageWorld extends SimpleGui {
         );
         setSlot(18, GuiElements.back(previousGui));
         setSlot(26,
-            guiElementBuilder(Items.TNT, "delete")
+            new GuiElementBuilder(Items.TNT)
+                .setName(Component.literal("Delete World").withStyle(ChatFormatting.RED, ChatFormatting.BOLD))
+                .addLoreLine(Component.literal("Deletes the world, kicking all players that are currently in it.").withStyle(ChatFormatting.GRAY))
                 .setCallback(() -> {
                     if (Permissions.check(player, "worldmanager.command.worldmanager.delete", 2)) {
                         new ConfirmTypeGui(player, id.toString(), () -> {
@@ -83,16 +89,5 @@ public class ManageWorld extends SimpleGui {
                     }
                 })
         );
-    }
-
-    private GuiElementBuilder guiElementBuilder(Item item, String id) {
-        return guiElementBuilder(new GuiElementBuilder(item), id);
-    }
-
-    private GuiElementBuilder guiElementBuilder(GuiElementBuilder builder, String id) {
-        return builder
-            .setName(LocalizedMessage.builder("worldmanager.gui.manage_world." + id + ".name")
-                .build())
-            .addLoreLine(LocalizedMessage.builder("worldmanager.gui.manage_world." + id + ".lore").build());
     }
 }